znemozneni rozsazení lidi z trida > 0

+++ Trida.cs
@@ -2,7 +2,7 @@
 {
     public class Trida
     {
-        private long id;
+        private int id;
         private string nazev;
         private int sirka;
         private int vyska;
@@ -10,7 +10,7 @@
         private string data;
         private List<int[]>[,] prijatelneKategorieMista;
 
-        public long Id { get { return id; } }
+        public int Id { get { return id; } }
         public string Nazev { get { return nazev; } set { nazev = value; } }
         public int Sirka { get { return sirka; } set { sirka = value; } }
         public int Vyska { get { return vyska; } set { vyska = value; } }
@@ -27,7 +27,7 @@
         /// <param name="vyska">Výška třídy (v polích)</param>
         /// <param name="rozsazena">Třída je/není rozsazená</param>
         /// <param name="data">Data rozsazené třídy</param>
-        public Trida(long id, string nazev, int sirka, int vyska, string data, bool jeRozsazena = false)
+        public Trida(int id, string nazev, int sirka, int vyska, string data, bool jeRozsazena = false)
         {
             this.id = id;
             this.nazev = nazev;
